[refactor] 파일 시스템 구현 개편
연관관계를 별도로 설정할 필요가 없게 수정, FileRepositoryDelegator 사용하지 않고 간단하게 구현
