--- src/main/java/com/prgrms/ohouse/domain/common/ImageAttachable.java
@@ -1,8 +1,0 @@
-package com.prgrms.ohouse.domain.common;
-
-/**
- * 이미지를 붙일 수 있는 타입임을 표시하는 Marker Interface 입니다.
- * 이미지 파일과 연관관계를 맺기 위해 인자 전달을 할 때 타입 체크 용으로만 사용됩니다.
- */
-public interface ImageAttachable {
-}
